third_party/WebKit/LayoutTests/inspector/console/console-eval-scoped-expected.txt

CONSOLE MESSAGE: line 4: with: Object property value
CONSOLE MESSAGE: line 1: eval in with: Object property value
CONSOLE MESSAGE: line 1: eval in function: Function parameter
Tests that evaluating 'console.log()' in the console will have access to its outer scope variables. Bug 60547.


Running: testSnippet1
VM:4 with: Object property value
VM:1 eval in with: Object property value

Running: testSnippet2
VM:1 eval in function: Function parameter

Running: testConsoleEvalPrimitiveValue
foo
"bar"

Running: testConsoleEvalObject
{
    configurable : true
    enumerable : true
    get : {
        className : "Object"
        description : "function getter() { }"
        objectId : <string>
        type : "function"
    }
    isOwn : true
    name : "getter"
    set : {
        type : "undefined"
    }
}
{
    configurable : true
    enumerable : true
    get : {
        type : "undefined"
    }
    isOwn : true
    name : "setter"
    set : {
        className : "Object"
        description : "function setter(_) { }"
        objectId : <string>
        type : "function"
    }
}
{
    configurable : true
    enumerable : true
    isOwn : true
    name : "baz"
    value : {
        type : "string"
        value : "baz"
    }
    writable : true
}
{
    configurable : true
    enumerable : false
    isOwn : true
    name : "__proto__"
    value : {
        className : "Object"
        description : "Object"
        objectId : <string>
        type : "object"
    }
    writable : true
}

Running: testGetEventListenersDoesNotThrow
getEventListeners(document.body.firstChild)
Object {}
